Created CardInfo component to render property information in a list
Added a new React component called "CardInfo" to render property information for each item in the JSON file.
Implemented the necessary code within the CardInfo component to map through the property data from the JSON file and render the information in a list format.
Utilized the CARD component (previously created) within the CardInfo component to display the property details.